// src/serverApi/v3/models/migration-body.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface MigrationBody
 */
export interface MigrationBody {
    /**
     * Set if migration is possible in this school
     * @type {boolean}
     * @memberof MigrationBody
     */
    oauthMigrationPossible?: boolean | null;
    /**
     * Set if migration is mandatory in this school
     * @type {boolean}
     * @memberof MigrationBody
     */
    oauthMigrationMandatory?: boolean | null;
    /**
     * Set if migration is finished in this school
     * @type {boolean}
     * @memberof MigrationBody
     */
    oauthMigrationFinished?: boolean | null;
}